atcoder/200-299/d/abc216_d.cpp

+++ atcoder/200-299/d/abc216_d.cpp
@@ -29,7 +29,7 @@ int main() {
 
     while (!to_parse.empty()) {
         vector<int> to_add{};
-        for (int i = 0; i < to_parse.size(); ++i) {
+        for (size_t i = 0; i < to_parse.size(); ++i) {
             if (!cylinder[i].empty()) {
                 value = cylinder[i].top();
